test(supabase): move Supabase unit tests into stack-supabase (#627)
Moves the 11 supabase* unit suites + the 2 test helpers (supabase-mock,
postgrest-wire) into the new package; rewrites @/ imports (including vi.mock and
dynamic import() targets) to relative ../src paths and @cipherstash/stack public
subpaths; repoints the v3 matrix catalog to @cipherstash/test-kit/catalog. Adds
fast-check + the @/ / wasm-inline aliases the source-resolved stack needs.

stack-supabase: 442 unit tests + type tests green. stack: 1172 pass (only the
pre-existing DATABASE_URL live suite fails locally).
