+++ packages/stack-supabase/__tests__/supabase-v3-wire.test.ts
@@ -9,8 +9,8 @@
  */
 
 import { describe, expect, it } from 'vitest'
-import { encryptedTable, types } from '@/eql/v3'
-import { EncryptedQueryBuilderV3Impl } from '@/supabase/query-builder-v3'
+import { encryptedTable, types } from '@cipherstash/stack/eql/v3'
+import { EncryptedQueryBuilderV3Impl } from '../src/query-builder-v3'
 import { createWirePostgrest } from './helpers/postgrest-wire'
 import { createMockEncryptionClient } from './helpers/supabase-mock'
