+++ packages/stack-supabase/__tests__/supabase-verify.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from 'vitest'
-import { encryptedTable, types } from '@/eql/v3'
-import type { IntrospectionResult } from '@/supabase/introspect'
-import { verifyDeclaredSchemas } from '@/supabase/verify'
+import { encryptedTable, types } from '@cipherstash/stack/eql/v3'
+import type { IntrospectionResult } from '../src/introspect'
+import { verifyDeclaredSchemas } from '../src/verify'
 
 const introspection: IntrospectionResult = [
   {
